fix #1485
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@11313 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ inc/db_functions.inc.php
@@ -135,7 +135,7 @@
       }
       
       // set cache  
-      $cache->set($id, $records , DB_CACHE_EXPIRE);
+      $cache->set($id, $records , (int)DB_CACHE_EXPIRE);
     } else {
       if (defined('STORE_DB_TRANSACTIONS') && STORE_DB_TRANSACTIONS == 'true') {
         $queryEndTime = array_sum(explode(" ",microtime()));
